+++ src/commonMain/kotlin/com/bkahlert/kommons/test/matchesGlob.kt
@@ -2,6 +2,7 @@ package com.bkahlert.kommons.test
 
 import com.bkahlert.kommons.test.com.bkahlert.kommons.LineSeparators
 import com.bkahlert.kommons.test.com.bkahlert.kommons.LineSeparators.isMultiline
+import com.bkahlert.kommons.test.com.bkahlert.kommons.matchesCurly
 import com.bkahlert.kommons.test.com.bkahlert.kommons.matchesGlob
 import com.bkahlert.kommons.test.com.bkahlert.kommons.toHexadecimalString
 import io.kotest.assertions.print.print
@@ -12,7 +13,10 @@ import io.kotest.matchers.should
 import io.kotest.matchers.shouldNot
 
 /**
- * Verifies that this character sequence matches the given glob [pattern].
+ * Verifies that this character sequence matches the given
+ * glob [pattern], using
+ * `*` to match within lines, and
+ * `**` to match across lines.
  *
  * **Example**
  * ```kotlin
@@ -37,7 +41,7 @@ import io.kotest.matchers.shouldNot
  * ```
  *
  * @see shouldNotMatchGlob
- * @see matchesGlob
+ * @see matchGlob
  */
 public infix fun <A : CharSequence> A?.shouldMatchGlob(pattern: String): A {
     this should matchGlob(pattern)
@@ -45,10 +49,13 @@ public infix fun <A : CharSequence> A?.shouldMatchGlob(pattern: String): A {
 }
 
 /**
- * Verifies that this character sequence does not match the given glob [pattern].
+ * Verifies that this character sequence does not match the given
+ * glob [pattern], using
+ * `*` to match within lines, and
+ * `**` to match across lines.
  *
  * @see shouldMatchGlob
- * @see matchesGlob
+ * @see matchGlob
  */
 public infix fun <A : CharSequence> A?.shouldNotMatchGlob(pattern: String): A {
     this shouldNot matchGlob(pattern)
@@ -56,7 +63,8 @@ public infix fun <A : CharSequence> A?.shouldNotMatchGlob(pattern: String): A {
 }
 
 /**
- * Verifies that this character sequence matches the given glob [pattern] using
+ * Verifies that this character sequence matches the given
+ * glob [pattern] using
  * the specified [wildcard] (default: `*`) to match within lines, and
  * the specified [multilineWildcard] (default: `**`) to match across lines.
  *
@@ -89,8 +97,8 @@ public infix fun <A : CharSequence> A?.shouldNotMatchGlob(pattern: String): A {
  * )
  * ```
  *
+ * @see shouldMatchGlob
  * @see shouldNotMatchGlob
- * @see matchesGlob
  */
 public fun matchGlob(
     pattern: CharSequence,
@@ -99,15 +107,121 @@ public fun matchGlob(
     vararg lineSeparators: String = LineSeparators.Common,
 ): Matcher<CharSequence?> = neverNullMatcher { value ->
     val description = buildString {
-        append("match the following glob pattern with")
+        append("match the following glob pattern")
         append(" (wildcard: $wildcard")
         append(", multiline wildcard: $multilineWildcard")
         append(", line separators: ${describeLineSeparators(*lineSeparators)}")
-        appendLine(")")
+        appendLine("):")
         append(describeString(pattern))
     }
     MatcherResult(
         value.matchesGlob(pattern, wildcard, multilineWildcard, *lineSeparators),
+        { describeString(value, "should", description) },
+        { describeString(value, "should not", description) },
+    )
+}
+
+
+/**
+ * Verifies that this character sequence matches the given
+ * SLF4J / Logback style [pattern], using
+ * `{}` to match within lines, and
+ * `{{}}` to match across lines.
+ *
+ * **Example**
+ * ```kotlin
+ * val multilineString = """
+ *     foo
+ *       .bar()
+ *       .baz()
+ * """.trimIndent()
+ *
+ * // ✅ matches thanks to the multiline wildcard **
+ * multilineString shouldMatchGlob """
+ *     foo
+ *       .{{}}()
+ * """.trimIndent()
+ *
+ * // ❌ fails to match since the simple wildcard *
+ * // does not match across line breaks
+ * multilineString shouldMatchGlob """
+ *     foo
+ *       .{}()
+ * """.trimIndent()
+ * ```
+ *
+ * @see shouldNotMatchCurly
+ * @see matchCurly
+ */
+public infix fun <A : CharSequence> A?.shouldMatchCurly(pattern: String): A {
+    this should matchCurly(pattern)
+    return this!!
+}
+
+/**
+ * Verifies that this character sequence does not match the given
+ * SLF4J / Logback style [pattern], using
+ * `{}` to match within lines, and
+ * `{{}}` to match across lines.
+ *
+ * @see shouldMatchCurly
+ * @see matchCurly
+ */
+public infix fun <A : CharSequence> A?.shouldNotMatchCurly(pattern: String): A {
+    this shouldNot matchCurly(pattern)
+    return this!!
+}
+
+/**
+ * Verifies that this character sequence matches the given
+ * SLF4J / Logback style [pattern], using
+ * `{}` to match within lines, and
+ * `{{}}` to match across lines.
+ *
+ * The specified [lineSeparators] (default: `\r\n`, `\n`, and `\r`) can be used
+ * interchangeably.
+ *
+ * **Example**
+ * ```kotlin
+ * val multilineString = """
+ *     foo
+ *       .bar()
+ *       .baz()
+ * """.trimIndent()
+ *
+ * // ✅ matches thanks to the multiline wildcard **
+ * multilineString should matchGlob(
+ * """
+ *     foo
+ *       .{{}}()
+ * """.trimIndent()
+ * )
+ *
+ * // ❌ fails to match since the simple wildcard *
+ * // does not match across line breaks
+ * multilineString should matchGlob(
+ * """
+ *     foo
+ *       .{}()
+ * """.trimIndent()
+ * )
+ * ```
+ *
+ * @see shouldMatchCurly
+ * @see shouldNotMatchCurly
+ */
+public fun matchCurly(
+    pattern: CharSequence,
+    vararg lineSeparators: String = LineSeparators.Common,
+): Matcher<CharSequence?> = neverNullMatcher { value ->
+    val description = buildString {
+        append("match the following curly pattern")
+        append(" (line separators: ${describeLineSeparators(*lineSeparators)}")
+        appendLine("):")
+        append(describeString(pattern))
+    }
+    MatcherResult(
+        value.matchesCurly(pattern, *lineSeparators),
         { describeString(value, "should", description) },
         { describeString(value, "should not", description) },
     )