+++ src/commonTest/kotlin/com/bkahlert/kommons/test/MatchesGlobKtTest.kt
@@ -3,6 +3,7 @@
 package com.bkahlert.kommons.test
 
 import com.bkahlert.kommons.test.com.bkahlert.kommons.LineSeparators
+import com.bkahlert.kommons.test.com.bkahlert.kommons.multilineGlobMatchInput
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.should
@@ -13,33 +14,29 @@ class MatchesGlobKtTest {
 
     @Test fun should_match_glob() {
         shouldNotThrowAny {
-            matchGlobInput shouldMatchGlob """
-                foo.*
-                bar[0]++
-                baz did **
+            multilineGlobMatchInput shouldMatchGlob """
+                foo
+                  .**()
             """.trimIndent()
         }
     }
 
     @Test fun should_match_glob__failure() {
         shouldThrow<AssertionError> {
-            matchGlobInput shouldMatchGlob """
-                foo.*
-                bar[0]++
-                baz did *
+            multilineGlobMatchInput shouldMatchGlob """
+                foo
+                  .*()
             """.trimIndent()
         }.message shouldBe """
             ""${'"'}
-            foo.bar()
-            bar[0]++
-            baz did throw a RuntimeException
-                at SomeFile.kt:42
+            foo
+              .bar()
+              .baz()
             ""${'"'}
-            should match the following glob pattern with (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r))
+            should match the following glob pattern (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r)):
             ""${'"'}
-            foo.*
-            bar[0]++
-            baz did *
+            foo
+              .*()
             ""${'"'}
         """.trimIndent()
     }
@@ -47,33 +44,29 @@ class MatchesGlobKtTest {
 
     @Test fun should_not_match_glob() {
         shouldNotThrowAny {
-            matchGlobInput shouldNotMatchGlob """
-                foo.*
-                bar[0]++
-                baz did *
+            multilineGlobMatchInput shouldNotMatchGlob """
+                foo
+                  .*()
             """.trimIndent()
         }
     }
 
     @Test fun should_not_match_glob__failure() {
         shouldThrow<AssertionError> {
-            matchGlobInput shouldNotMatchGlob """
-                foo.*
-                bar[0]++
-                baz did **
+            multilineGlobMatchInput shouldNotMatchGlob """
+                foo
+                  .**()
             """.trimIndent()
         }.message shouldBe """
             ""${'"'}
-            foo.bar()
-            bar[0]++
-            baz did throw a RuntimeException
-                at SomeFile.kt:42
+            foo
+              .bar()
+              .baz()
             ""${'"'}
-            should not match the following glob pattern with (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r))
+            should not match the following glob pattern (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r)):
             ""${'"'}
-            foo.*
-            bar[0]++
-            baz did **
+            foo
+              .**()
             ""${'"'}
         """.trimIndent()
     }
@@ -81,11 +74,10 @@ class MatchesGlobKtTest {
 
     @Test fun match_glob() {
         shouldNotThrowAny {
-            matchGlobInput should matchGlob(
+            multilineGlobMatchInput should matchGlob(
                 """
-                foo.{}
-                bar[0]++
-                baz did {{}}
+                foo
+                {}.{{}}()
             """.trimIndent(),
                 wildcard = "{}",
                 multilineWildcard = "{{}}",
@@ -97,11 +89,10 @@ class MatchesGlobKtTest {
     @Test fun match_glob__failure() {
         @Suppress("LongLine")
         shouldThrow<AssertionError> {
-            matchGlobInput should matchGlob(
+            multilineGlobMatchInput should matchGlob(
                 """
-                foo.{}
-                bar[0]++
-                baz did {}
+                foo
+                {}.{}()
             """.trimIndent(),
                 wildcard = "{}",
                 multilineWildcard = "{{}}",
@@ -109,16 +100,14 @@ class MatchesGlobKtTest {
             )
         }.message shouldBe """
             ""${'"'}
-            foo.bar()
-            bar[0]++
-            baz did throw a RuntimeException
-                at SomeFile.kt:42
+            foo
+              .bar()
+              .baz()
             ""${'"'}
-            should match the following glob pattern with (wildcard: {}, multiline wildcard: {{}}, line separators: CRLF (\r\n), LF (\n), CR (\r), NEL (\u0085), PS (\u2029), LS (\u2028), Unknown (0xf09faba0))
+            should match the following glob pattern (wildcard: {}, multiline wildcard: {{}}, line separators: CRLF (\r\n), LF (\n), CR (\r), NEL (\u0085), PS (\u2029), LS (\u2028), Unknown (0xf09faba0)):
             ""${'"'}
-            foo.{}
-            bar[0]++
-            baz did {}
+            foo
+            {}.{}()
             ""${'"'}
         """.trimIndent()
     }
@@ -129,15 +118,116 @@ class MatchesGlobKtTest {
             "foo.bar()" should matchGlob("bar.{}")
         }.message shouldBe """
             "foo.bar()"
-            should match the following glob pattern with (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r))
+            should match the following glob pattern (wildcard: *, multiline wildcard: **, line separators: CRLF (\r\n), LF (\n), CR (\r)):
+            "bar.{}"
+        """.trimIndent()
+    }
+
+
+    @Test fun should_match_curly() {
+        shouldNotThrowAny {
+            multilineGlobMatchInput shouldMatchCurly """
+                foo
+                  .{{}}()
+            """.trimIndent()
+        }
+    }
+
+    @Test fun should_match_curly__failure() {
+        shouldThrow<AssertionError> {
+            multilineGlobMatchInput shouldMatchCurly """
+                foo
+                  .{}()
+            """.trimIndent()
+        }.message shouldBe """
+            ""${'"'}
+            foo
+              .bar()
+              .baz()
+            ""${'"'}
+            should match the following curly pattern (line separators: CRLF (\r\n), LF (\n), CR (\r)):
+            ""${'"'}
+            foo
+              .{}()
+            ""${'"'}
+        """.trimIndent()
+    }
+
+
+    @Test fun should_not_match_curly() {
+        shouldNotThrowAny {
+            multilineGlobMatchInput shouldNotMatchCurly """
+                foo
+                  .{}()
+            """.trimIndent()
+        }
+    }
+
+    @Test fun should_not_match_curly__failure() {
+        shouldThrow<AssertionError> {
+            multilineGlobMatchInput shouldNotMatchCurly """
+                foo
+                  .{{}}()
+            """.trimIndent()
+        }.message shouldBe """
+            ""${'"'}
+            foo
+              .bar()
+              .baz()
+            ""${'"'}
+            should not match the following curly pattern (line separators: CRLF (\r\n), LF (\n), CR (\r)):
+            ""${'"'}
+            foo
+              .{{}}()
+            ""${'"'}
+        """.trimIndent()
+    }
+
+
+    @Test fun match_curly() {
+        shouldNotThrowAny {
+            multilineGlobMatchInput should matchCurly(
+                """
+                foo
+                {}.{{}}()
+            """.trimIndent(),
+                *LineSeparators.Unicode, "🫠",
+            )
+        }
+    }
+
+    @Test fun match_curly__failure() {
+        @Suppress("LongLine")
+        shouldThrow<AssertionError> {
+            multilineGlobMatchInput should matchCurly(
+                """
+                foo
+                {}.{}()
+            """.trimIndent(),
+                *LineSeparators.Unicode, "🫠",
+            )
+        }.message shouldBe """
+            ""${'"'}
+            foo
+              .bar()
+              .baz()
+            ""${'"'}
+            should match the following curly pattern (line separators: CRLF (\r\n), LF (\n), CR (\r), NEL (\u0085), PS (\u2029), LS (\u2028), Unknown (0xf09faba0)):
+            ""${'"'}
+            foo
+            {}.{}()
+            ""${'"'}
+        """.trimIndent()
+    }
+
+    @Test fun match_curly__singleLine_failure() {
+        @Suppress("LongLine")
+        shouldThrow<AssertionError> {
+            "foo.bar()" should matchCurly("bar.{}")
+        }.message shouldBe """
+            "foo.bar()"
+            should match the following curly pattern (line separators: CRLF (\r\n), LF (\n), CR (\r)):
             "bar.{}"
         """.trimIndent()
     }
 }
-
-internal val matchGlobInput = """
-    foo.bar()
-    bar[0]++
-    baz did throw a RuntimeException
-        at SomeFile.kt:42
-""".trimIndent()