+++ src/jvmMain/kotlin/com/bkahlert/kommons/test/com/bkahlert/kommons/debug/rendering.kt
@@ -1,18 +1,10 @@
 package com.bkahlert.kommons.test.com.bkahlert.kommons.debug
 
+import com.bkahlert.kommons.test.com.bkahlert.kommons.LineSeparators
 import com.bkahlert.kommons.test.com.bkahlert.kommons.quoted
 import com.bkahlert.kommons.test.rootCause
 import java.nio.file.Path
 
-
-internal val LineSeparators: Array<String> = arrayOf(
-    "\u000D\u000A",
-    "\u000A",
-    "\u000D",
-    "\u0085",
-    "\u2029",
-    "\u2028",
-)
 
 internal fun Any?.toCompactString(): String {
     return when (this) {
@@ -21,13 +13,13 @@ internal fun Any?.toCompactString(): String {
         is Array<*> -> toList().toCompactString()
         is Iterable<*> -> joinToString(prefix = "[", postfix = "]") { it.toCompactString() }
         is Process -> also { waitFor() }.exitValue().toString()
-        is CharSequence -> split(*LineSeparators).joinToString(separator = "⏎").removeSuffix("⏎").quoted
+        is CharSequence -> split(*LineSeparators.Unicode).joinToString(separator = "⏎").removeSuffix("⏎").quoted
         else -> when (this) {
             null -> "null"
             Unit -> ""
             else -> {
                 val string = toCustomStringOrNull() ?: renderType(simplified = true)
-                string.split(*LineSeparators).joinToString(separator = "⏎").removeSuffix("⏎")
+                string.split(*LineSeparators.Unicode).joinToString(separator = "⏎").removeSuffix("⏎")
             }
         }
     }